Simplon_FastAPI/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
catboost==1.2.7
cffi==1.17.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.11
fonttools==4.56.0
graphviz==0.20.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.29.0
numba==0.61.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
plotly==6.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pyodbc==5.2.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-tds==1.16.0
pytz==2025.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
shap==0.47.0
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlalchemy_pytds==1.0.2
sqlmodel==0.0.23
starlette==0.46.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0